5 song playback from libraryfragment (#7)
* Align 'SelectedPlaylistHolder' & 'SelectedSongHolder' to only hold 'mainActivity' variable

No need to store 'spotifyApi' & 'spotifyAppRemote' as they can be accessed from the 'mainActivity' variable

* Clean 'LibraryFragment'

* Fix Song Playback across both 'LibraryFragment' and 'PlaylistView'

The code for 'SongView' had been constructed around clicking on a song from a playlist and so, doing the same in the library fragment caused the app to crash. The fix I have here is to state where 'SongView' is being called from and then obtain the required 'mainActivity' in the corresponding way.
For example, 'mainActivity' could not be found from 'LibraryFragment' when clicking on a song in a playlist and vice versa.