Improve playback handling of snippets
If requesting to play a snippet of the song that is currently being played, reduce the amount of time needed to play the snippet.

If the song is currently playing, just seek to the start of the snippet.
    If paused, seek and then play.
If the requested snippet is from a song not currently being played, continue normal behaviour: play start of selected snippet's song, immediately pause, seek to start of snippet, then play and set timer to stop playback at end of snippet.