9 update songview UI (#20)
* Update Play button

* Add fragments for song notes and snippets

* Add Adapter for notes/snippets tabs on `SongView`

* Clean imports

* Update `activity_song_view.xml`

* Improve AppBarLayout:
Only show AppBar when scrolled to top of scrollable elements (view_pager2);
Add indicator for song details;
Adjust positioning

* Add custom tab layouts for `SongView`

* Add logic for custom tabs and count-badges

The count-badges displays the amount of notes/snippets created for a song

* Improve Notes UI

Press a button to add new notes.

* Add Snippets UI

* Restyle and reposition play button on `SongView`

* Improve spacing for `SongView` tabs

* Improve playback handling of snippets

If requesting to play a snippet of the song that is currently being played, reduce the amount of time needed to play the snippet.

If the song is currently playing, just seek to the start of the snippet.
    If paused, seek and then play.
If the requested snippet is from a song not currently being played, continue normal behaviour: play start of selected snippet's song, immediately pause, seek to start of snippet, then play and set timer to stop playback at end of snippet.