38 playback device checker (#39)
* Add `DeviceChecker` utility class

This utility class checks if the currently active Spotify playback device matches a configured device name.

- It uses SharedPreferences to store the enabled state and the target device name.
- Provides a callback interface for handling the check result and warning state changes.
- The check is performed asynchronously and results are posted back to the main thread.
- Includes a method to check if the device check feature is enabled.

* Add user-read-playback-state scope to Spotify authentication
Hide the "Play from Earliest Listened To" button for now, until functionality implemented for it

* Perform initial device check on app startup

This commit introduces an initial device check when the MainActivity starts.
It uses the `DeviceChecker` utility to verify if the correct playback device is selected and displays a warning toast message if not, provided that the device check feature is enabled.

* Integrate playback device check into `PlaybackManager`

The `checkPlaybackDevice` method is now called:
- When the track changes.
- When a new song starts playing.

This ensures that the application checks if playback is happening on the correct device

* Add playback device warning indicator

This commit introduces a visual indicator to alert the user when playback is occurring on a device other than the current one.

- A warning icon is added to the playback bar UI.
- `PlaybackManager` now listens for `onDeviceWarningStateChanged` and notifies `BaseActivity` instances

* Add "Device Check" settings

This commit introduces a new "Device Check" section in the settings menu. Users can now:

- Enable or disable checking the playback device.
- Specify their device name.
- Discover available Spotify devices on their network.

The settings are persisted using SharedPreferences. The UI for device name input, discovery button, and available devices list is conditionally shown based on whether the "Check playback device" option is enabled.

* Fix: Ensure device warning icon is displayed across all activities

When a new Activity is registered as a listener, update its
`DeviceWarning` to match the current state.

Previously, `DeviceWarning` was only displayed on currently active activities.

* Do not record listen if not playing on user's device