Check for song in cache when selecting playback-bar
Relevant song details would never be provided when clicking on the playback-bar to launch `SongView`. Now, a check will be made to the `SongCache` for the relevant `SongModel` with all information present. This `SongModel` should exist if a playlist containing it has been cached previously.

TODO: If the song does not exist in the cache, there is no current solution in place. What should happen is an async call to fetch the relevant details. When all information is present then, launch `SongView` if it didn't take too long. Finally, cache the fully-detailed song.